Revamped base in its own module
